requirements.txt
mimesis == 14.0.0
pydantic == 1.10.8
pytest == 8.0.1
requests == 2.31.0
allure-pytest == 2.13.2
curlify == 2.2.1